implimented main code using switch case

+++ Testing-kit/Testing-kit/L293D_PWM_Library.c
@@ -1,4 +1,4 @@
-#include "PWM_Library.h"
+#include "L293D_PWM_Library.h"
 
 void InitPWM(void)
 {
@@ -9,20 +9,17 @@ void InitPWM(void)
 	
 	 
 	TCCR3A = (1<<PwmType1 |1<<PwmType2 | 1<<WGM31);			//fast pwm mode
-	TCCR3B =(1<<WGM33 | 1<<WGM32 | 1<<CS32);    //prescaller 256
+	TCCR3B =(1<<WGM33 | 1<<WGM32 | 1<<CS32);				//prescaller 256
 	
 	
 }
 
-void SetICR(float n)
+void SetICR(int n)
 {
 	ICR3= (n*31250)/1000;
 }
 
-void SetDutyCycle(float n)
+void SetDutyCycle(int n)
 {
-	
-	
-
 		OCR3A = ICR3 - (n*31250)/1000;
 }
